+++ src/credentials/dynamic-sha2.ts
